+++ python/probros/translators/gen_translator.py
@@ -1,7 +1,19 @@
 import ast
 from julia_translator import JuliaTranslator
 
-class TuringTranslator(JuliaTranslator):
+class GenTranslator(JuliaTranslator):
+    def visit_Call(self, node):
+        match node:
+            case ast.Call(
+                func=ast.Attribute(value=ast.Name(id="pr"),attr="IndexedAddress"),
+                args = [ast.Constant(value=address), index]
+                ):
+                if isinstance(index, ast.Name):
+                    self.write(":" + address + " => " + index.id)
+                    return
+
+        super().visit_Call(node)
+
     def probprog(self, name: str, args: list[str], body):
         self.write("using Turing\n")
         self.write("using Distributions\n")
@@ -10,19 +22,20 @@ class TuringTranslator(JuliaTranslator):
             self.traverse(body)
 
     def probprog_sample(self, target, address, distribution_name, distribution_args, distribution_keywords):
-        # TODO: fail if indexedaddress but not indexed target
         self.fill()
         self.set_precedence(ast._Precedence.TUPLE, target)
         self.traverse(target)
+        self.write(" = ")
+        with self.delimit("{", "}"):
+            self.traverse(address)
         self.write(" ~ ")
-        #self.traverse(address)
         self.write(f"{distribution_name}")
         self._write_arguments(distribution_args, distribution_keywords)
 
     def probprog_observe(self, value, address, distribution_name, distribution_args, distribution_keywords):
-        self.traverse(value)
+        with self.delimit("{", "}"):
+            self.traverse(address)
         self.write(" ~ ")
-        # self.traverse(address)
         self.write(f"{distribution_name}")
         self._write_arguments(distribution_args, distribution_keywords)
     
@@ -31,4 +44,3 @@ class TuringTranslator(JuliaTranslator):
     
     def probprog_factor(self, value):
         raise NotImplementedError
-
