+++ python/probros/translators/translator.py
@@ -1,13 +1,10 @@
 
 # %%
 # https://github.com/python/cpython/blob/3.12/Lib/ast.py
-from _ast import FunctionDef
 import ast
 from contextlib import contextmanager, nullcontext
 from abc import ABC, abstractmethod
 from typing import Any
-import sys
-_INFSTR = "1e" + repr(sys.float_info.max_10_exp + 1)
 
 class Translator(ast.NodeVisitor, ABC):
 
@@ -146,7 +143,7 @@ class Translator(ast.NodeVisitor, ABC):
     
 
     # Probabilistic Program entry point
-    def visit_FunctionDef(self, node: FunctionDef) -> Any:
+    def visit_FunctionDef(self, node: ast.FunctionDef) -> Any:
         match node:
             case ast.FunctionDef(
                 name=probprog_name,
@@ -239,711 +236,3 @@ class Translator(ast.NodeVisitor, ABC):
     @abstractmethod
     def probprog_factor(self, value):
         raise NotImplementedError
-
-class PythonTranslator(Translator):
-    def __init__(self):
-        super().__init__()
-        self._indent = 0
-
-    @contextmanager
-    def block(self):
-        """A context manager for preparing the source for blocks. It adds
-        the character':', increases the indentation on enter and decreases
-        the indentation on exit.
-        """
-        self.write(":")
-        self._indent += 1
-        yield
-        self._indent -= 1
-
-    def visit_Expr(self, node):
-        self.fill()
-        self.set_precedence(ast._Precedence.YIELD, node.value)
-        self.traverse(node.value)
-    
-    def visit_Break(self, node):
-        self.fill("break")
-
-    def visit_Assign(self, node):
-        self.fill()
-        for target in node.targets:
-            self.set_precedence(ast._Precedence.TUPLE, target)
-            self.traverse(target)
-            self.write(" = ")
-        self.traverse(node.value)
-
-    def visit_For(self, node):
-        self.fill("for ")
-        self.set_precedence(ast._Precedence.TUPLE, node.target)
-        self.traverse(node.target)
-        self.write(" in ")
-        self.traverse(node.iter)
-        with self.block():
-            self.traverse(node.body)
-
-    def visit_If(self, node):
-        self.fill("if ")
-        self.traverse(node.test)
-        with self.block():
-            self.traverse(node.body)
-        # collapse nested ifs into equivalent elifs.
-        while node.orelse and len(node.orelse) == 1 and isinstance(node.orelse[0], ast.If):
-            node = node.orelse[0]
-            self.fill("elif ")
-            self.traverse(node.test)
-            with self.block():
-                self.traverse(node.body)
-        # final else
-        if node.orelse:
-            self.fill("else")
-            with self.block():
-                self.traverse(node.orelse)
-
-    def visit_While(self, node):
-        self.fill("while ")
-        self.traverse(node.test)
-        with self.block():
-            self.traverse(node.body)
-
-    def visit_Name(self, node):
-        self.write(node.id)
-
-    def _write_constant(self, value):
-        if isinstance(value, (float, complex)):
-            # Substitute overflowing decimal literal for AST infinities,
-            # and inf - inf for NaNs.
-            # should not occur in practice
-            self.write(
-                repr(value)
-                .replace("inf", _INFSTR)
-                .replace("nan", f"({_INFSTR}-{_INFSTR})")
-            )
-        else:
-            self.write(repr(value))
-
-    def visit_Constant(self, node):
-        value = node.value
-        if isinstance(value, tuple):
-            with self.delimit("(", ")"):
-                self.items_view(self._write_constant, value)
-        elif value is ...:
-            self.write("...")
-        else:
-            if node.kind == "u":
-                self.write("u")
-            self._write_constant(node.value)
-
-    def visit_List(self, node):
-        with self.delimit("[", "]"):
-            self.interleave(lambda: self.write(", "), self.traverse, node.elts)
-
-    # value if test else other
-    #def visit_IfExp(self, node):
-    #    with self.require_parens(ast._Precedence.TEST, node):
-    #        self.set_precedence(ast._Precedence.TEST.next(), node.body, node.test)
-    #        self.traverse(node.body)
-    #        self.write(" if ")
-    #        self.traverse(node.test)
-    #        self.write(" else ")
-    #        self.set_precedence(ast._Precedence.TEST, node.orelse)
-    #        self.traverse(node.orelse)
-        
-    def visit_Tuple(self, node):
-        with self.delimit_if(
-            "(",
-            ")",
-            len(node.elts) == 0 or self.get_precedence(node) > ast._Precedence.TUPLE
-        ):
-            self.items_view(self.traverse, node.elts)
-
-
-    unop = {"Invert": "~", "Not": "not", "UAdd": "+", "USub": "-"}
-    unop_precedence = {
-        "not": ast._Precedence.NOT,
-        "~": ast._Precedence.FACTOR,
-        "+": ast._Precedence.FACTOR,
-        "-": ast._Precedence.FACTOR,
-    }
-
-    def visit_UnaryOp(self, node):
-        operator = self.unop[node.op.__class__.__name__]
-        operator_precedence = self.unop_precedence[operator]
-        with self.require_parens(operator_precedence, node):
-            self.write(operator)
-            # factor prefixes (+, -, ~) shouldn't be separated
-            # from the value they belong, (e.g: +1 instead of + 1)
-            if operator_precedence is not ast._Precedence.FACTOR:
-                self.write(" ")
-            self.set_precedence(operator_precedence, node.operand)
-            self.traverse(node.operand)
-
-    binop = {
-        "Add": "+",
-        "Sub": "-",
-        "Mult": "*",
-        "MatMult": "@",
-        "Div": "/",
-        "Mod": "%",
-        "LShift": "<<",
-        "RShift": ">>",
-        "BitOr": "|",
-        "BitXor": "^",
-        "BitAnd": "&",
-        "FloorDiv": "//",
-        "Pow": "**",
-    }
-
-    binop_precedence = {
-        "+": ast._Precedence.ARITH,
-        "-": ast._Precedence.ARITH,
-        "*": ast._Precedence.TERM,
-        "@": ast._Precedence.TERM,
-        "/": ast._Precedence.TERM,
-        "%": ast._Precedence.TERM,
-        "<<": ast._Precedence.SHIFT,
-        ">>": ast._Precedence.SHIFT,
-        "|": ast._Precedence.BOR,
-        "^": ast._Precedence.BXOR,
-        "&": ast._Precedence.BAND,
-        "//": ast._Precedence.TERM,
-        "**": ast._Precedence.POWER,
-    }
-
-    binop_rassoc = frozenset(("**",))
-    def visit_BinOp(self, node):
-        operator = self.binop[node.op.__class__.__name__]
-        operator_precedence = self.binop_precedence[operator]
-        with self.require_parens(operator_precedence, node):
-            if operator in self.binop_rassoc:
-                left_precedence = operator_precedence.next()
-                right_precedence = operator_precedence
-            else:
-                left_precedence = operator_precedence
-                right_precedence = operator_precedence.next()
-
-            self.set_precedence(left_precedence, node.left)
-            self.traverse(node.left)
-            self.write(f" {operator} ")
-            self.set_precedence(right_precedence, node.right)
-            self.traverse(node.right)
-
-    cmpops = {
-        "Eq": "==",
-        "NotEq": "!=",
-        "Lt": "<",
-        "LtE": "<=",
-        "Gt": ">",
-        "GtE": ">=",
-        "Is": "is",
-        "IsNot": "is not",
-        "In": "in",
-        "NotIn": "not in",
-    }
-
-    def visit_Compare(self, node):
-        with self.require_parens(ast._Precedence.CMP, node):
-            self.set_precedence(ast._Precedence.CMP.next(), node.left, *node.comparators)
-            self.traverse(node.left)
-            for o, e in zip(node.ops, node.comparators):
-                self.write(" " + self.cmpops[o.__class__.__name__] + " ")
-                self.traverse(e)
-
-    boolops = {"And": "and", "Or": "or"}
-    boolop_precedence = {"and": ast._Precedence.AND, "or": ast._Precedence.OR}
-
-    def visit_BoolOp(self, node):
-        operator = self.boolops[node.op.__class__.__name__]
-        operator_precedence = self.boolop_precedence[operator]
-
-        def increasing_level_traverse(node):
-            nonlocal operator_precedence
-            operator_precedence = operator_precedence.next()
-            self.set_precedence(operator_precedence, node)
-            self.traverse(node)
-
-        with self.require_parens(operator_precedence, node):
-            s = f" {operator} "
-            self.interleave(lambda: self.write(s), increasing_level_traverse, node.values)
-
-    def visit_Call(self, node):
-        self.set_precedence(ast._Precedence.ATOM, node.func)
-        self.traverse(node.func)
-        with self.delimit("(", ")"):
-            comma = False
-            for e in node.args:
-                if comma:
-                    self.write(", ")
-                else:
-                    comma = True
-                self.traverse(e)
-            for e in node.keywords:
-                if comma:
-                    self.write(", ")
-                else:
-                    comma = True
-                self.traverse(e)
-
-    def visit_Subscript(self, node):
-        def is_non_empty_tuple(slice_value):
-            return (
-                isinstance(slice_value, tuple)
-                and slice_value.elts
-            )
-
-        self.set_precedence(ast._Precedence.ATOM, node.value)
-        self.traverse(node.value)
-        with self.delimit("[", "]"):
-            if is_non_empty_tuple(node.slice):
-                # parentheses can be omitted if the tuple isn't empty
-                self.items_view(self.traverse, node.slice.elts)
-            else:
-                self.traverse(node.slice)
-
-    def visit_Attribute(self, node):
-        self.set_precedence(ast._Precedence.ATOM, node.value)
-        self.traverse(node.value)
-        if isinstance(node.value, ast.Constant) and isinstance(node.value.value, int):
-            self.write(" ")
-        self.write(".")
-        self.write(node.attr)
-
-    def visit_Return(self, node):
-        self.fill("return")
-        if node.value:
-            self.write(" ")
-            self.traverse(node.value)
-
-
-class PyroTranslator(PythonTranslator):
-    def visit_Call(self, node):
-        match node:
-            case ast.Call(
-                func=ast.Attribute(value=ast.Name(id="pr"),attr="IndexedAddress"),
-                args = [ast.Constant(value=address), index]
-                ):
-                if isinstance(index, ast.Name):
-                    self.write("f'" + address + "_{" + index.id + "}'")
-                    return
-
-        super().visit_Call(node)
-
-    def probprog(self, name: str, args: list[str], body):
-        self.write("import pyro\n")
-        self.write("import pyro.distributions as dist\n")
-        self.write(f"def {name}(", ", ".join(args), ")")
-        with self.block():
-            self.traverse(body)
-
-    def probprog_sample(self, target, address, distribution_name, distribution_args, distribution_keywords):
-        self.fill()
-        self.set_precedence(ast._Precedence.TUPLE, target)
-        self.traverse(target)
-        self.write(" = pyro.sample(")
-        self.traverse(address)
-        self.write(f", dist.{distribution_name}")
-        self._write_arguments(distribution_args, distribution_keywords)
-        self.write(")")
-
-    def probprog_observe(self, value, address, distribution_name, distribution_args, distribution_keywords):
-        self.write("pyro.sample(")
-        self.traverse(address)
-        self.write(f", dist.{distribution_name}")
-        self._write_arguments(distribution_args, distribution_keywords)
-        self.write(", observed=")
-        self.traverse(value)
-        self.write(")")
-    
-    def probprog_boolean_observe(self, value, address):
-        raise NotImplementedError
-    
-    def probprog_factor(self, value):
-        raise NotImplementedError
-
-
-class JuliaTranslator(Translator):
-    def __init__(self):
-        super().__init__()
-        self._indent = 0
-
-    @contextmanager
-    def block(self, write_end=True):
-        """A context manager for preparing the source for blocks. It adds
-        the character':', increases the indentation on enter and decreases
-        the indentation on exit.
-        """
-        self._indent += 1
-        yield
-        self._indent -= 1
-        if write_end:
-            self.fill("end")
-
-    def visit_Expr(self, node):
-        self.fill()
-        self.set_precedence(ast._Precedence.YIELD, node.value)
-        self.traverse(node.value)
-    
-    def visit_Break(self, node):
-        self.fill("break")
-
-    def visit_Assign(self, node):
-        self.fill()
-        for target in node.targets:
-            self.set_precedence(ast._Precedence.TUPLE, target)
-            self.traverse(target)
-            self.write(" = ")
-        self.traverse(node.value)
-
-    def visit_For(self, node):
-        self.fill("for ")
-        self.set_precedence(ast._Precedence.TUPLE, node.target)
-        self.traverse(node.target)
-        self.write(" in ")
-        match node.iter:
-            case ast.Call(
-                    func=ast.Name(id="range"),
-                    args=[end_iter]
-                ):
-                    self.write("1:")
-                    self.traverse(end_iter)
-                    self.write("-1")
-            case ast.Call(
-                    func=ast.Name(id="range"),
-                    args=[start_iter, end_iter]
-                ):
-                    self.traverse(start_iter)
-                    self.write(":")
-                    self.traverse(end_iter)
-                    self.write("-1")
-            case _:
-                raise Exception(f"Unsupported for loop range {ast.dump(node.iter)}")
-        with self.block():
-            self.traverse(node.body)
-
-    def visit_If(self, node):
-        self.fill("if ")
-        self.traverse(node.test)
-        with self.block(write_end=node.orelse is None):
-            self.traverse(node.body)
-        # collapse nested ifs into equivalent elifs.
-        while node.orelse and len(node.orelse) == 1 and isinstance(node.orelse[0], ast.If):
-            node = node.orelse[0]
-            self.fill("elseif ")
-            self.traverse(node.test)
-            with self.block(write_end = not (node.orelse and len(node.orelse) == 1 and isinstance(node.orelse[0], ast.If))):
-                self.traverse(node.body)
-        # final else
-        if node.orelse:
-            self.fill("else")
-            with self.block():
-                self.traverse(node.orelse)
-
-    def visit_While(self, node):
-        self.fill("while ")
-        self.traverse(node.test)
-        with self.block():
-            self.traverse(node.body)
-
-    def visit_Name(self, node):
-        self.write(node.id)
-
-    def _write_constant(self, value):
-        if isinstance(value, (float, complex)):
-            # Substitute overflowing decimal literal for AST infinities,
-            # and inf - inf for NaNs.
-            # should not occur in practice
-            self.write(
-                repr(value)
-                .replace("inf", _INFSTR)
-                .replace("nan", f"({_INFSTR}-{_INFSTR})")
-            )
-        elif isinstance(value, str):
-            self.write('"' + repr(value)[1:-1] + '"')
-        else:
-            self.write(repr(value))
-
-    def visit_Constant(self, node):
-        value = node.value
-        if isinstance(value, tuple):
-            with self.delimit("(", ")"):
-                self.items_view(self._write_constant, value)
-        elif value is ...:
-            self.write("...")
-        else:
-            if node.kind == "u":
-                self.write("u")
-            self._write_constant(node.value)
-
-    def visit_List(self, node):
-        with self.delimit("[", "]"):
-            self.interleave(lambda: self.write(", "), self.traverse, node.elts)
-
-    # value if test else other
-    #def visit_IfExp(self, node):
-    #    with self.require_parens(ast._Precedence.TEST, node):
-    #        self.set_precedence(ast._Precedence.TEST.next(), node.body, node.test)
-    #        self.traverse(node.body)
-    #        self.write(" if ")
-    #        self.traverse(node.test)
-    #        self.write(" else ")
-    #        self.set_precedence(ast._Precedence.TEST, node.orelse)
-    #        self.traverse(node.orelse)
-        
-    def visit_Tuple(self, node):
-        with self.delimit_if(
-            "(",
-            ")",
-            len(node.elts) == 0 or self.get_precedence(node) > ast._Precedence.TUPLE
-        ):
-            self.items_view(self.traverse, node.elts)
-
-
-    unop = {"Invert": "~", "Not": "!", "UAdd": "+", "USub": "-"}
-    unop_precedence = {
-        "not": ast._Precedence.NOT,
-        "~": ast._Precedence.FACTOR,
-        "+": ast._Precedence.FACTOR,
-        "-": ast._Precedence.FACTOR,
-    }
-
-    def visit_UnaryOp(self, node):
-        operator = self.unop[node.op.__class__.__name__]
-        operator_precedence = self.unop_precedence[operator]
-        with self.require_parens(operator_precedence, node):
-            self.write(operator)
-            # factor prefixes (+, -, ~) shouldn't be separated
-            # from the value they belong, (e.g: +1 instead of + 1)
-            if operator_precedence is not ast._Precedence.FACTOR:
-                self.write(" ")
-            self.set_precedence(operator_precedence, node.operand)
-            self.traverse(node.operand)
-
-    binop = {
-        "Add": "+",
-        "Sub": "-",
-        "Mult": "*",
-        "Div": "/",
-        "Mod": "%",
-        "BitOr": "|",
-        "BitXor": "^",
-        "BitAnd": "&",
-        "Pow": "^",
-    }
-
-    # TODO: map name not symbol
-    binop_precedence = {
-        "+": ast._Precedence.ARITH,
-        "-": ast._Precedence.ARITH,
-        "*": ast._Precedence.TERM,
-        "/": ast._Precedence.TERM,
-        "%": ast._Precedence.TERM,
-        "|": ast._Precedence.BOR,
-        "^": ast._Precedence.BXOR,
-        "&": ast._Precedence.BAND,
-        "^": ast._Precedence.POWER,
-    }
-
-    binop_rassoc = frozenset(("^",))
-    def visit_BinOp(self, node):
-        operator = self.binop[node.op.__class__.__name__]
-        operator_precedence = self.binop_precedence[operator]
-        with self.require_parens(operator_precedence, node):
-            if operator in self.binop_rassoc:
-                left_precedence = operator_precedence.next()
-                right_precedence = operator_precedence
-            else:
-                left_precedence = operator_precedence
-                right_precedence = operator_precedence.next()
-
-            self.set_precedence(left_precedence, node.left)
-            self.traverse(node.left)
-            self.write(f" {operator} ")
-            self.set_precedence(right_precedence, node.right)
-            self.traverse(node.right)
-
-    cmpops = {
-        "Eq": "==",
-        "NotEq": "!=",
-        "Lt": "<",
-        "LtE": "<=",
-        "Gt": ">",
-        "GtE": ">=",
-    }
-
-    def visit_Compare(self, node):
-        with self.require_parens(ast._Precedence.CMP, node):
-            self.set_precedence(ast._Precedence.CMP.next(), node.left, *node.comparators)
-            self.traverse(node.left)
-            for o, e in zip(node.ops, node.comparators):
-                self.write(" " + self.cmpops[o.__class__.__name__] + " ")
-                self.traverse(e)
-
-    boolops = {"And": "&&", "Or": "||"}
-    boolop_precedence = {"&&": ast._Precedence.AND, "||": ast._Precedence.OR}
-
-    def visit_BoolOp(self, node):
-        operator = self.boolops[node.op.__class__.__name__]
-        operator_precedence = self.boolop_precedence[operator]
-
-        def increasing_level_traverse(node):
-            nonlocal operator_precedence
-            operator_precedence = operator_precedence.next()
-            self.set_precedence(operator_precedence, node)
-            self.traverse(node)
-
-        with self.require_parens(operator_precedence, node):
-            s = f" {operator} "
-            self.interleave(lambda: self.write(s), increasing_level_traverse, node.values)
-
-    def visit_Call(self, node):
-        self.set_precedence(ast._Precedence.ATOM, node.func)
-        match node.func:
-            case ast.Name(id=func_name):
-                if func_name == "len":
-                    self.write("length")
-                else:
-                    self.write(func_name)
-            case _:
-                self.traverse(node.func)
-        with self.delimit("(", ")"):
-            comma = False
-            for e in node.args:
-                if comma:
-                    self.write(", ")
-                else:
-                    comma = True
-                self.traverse(e)
-            for e in node.keywords:
-                if comma:
-                    self.write(", ")
-                else:
-                    comma = True
-                self.traverse(e)
-
-    def visit_Subscript(self, node):
-        def is_non_empty_tuple(slice_value):
-            return (
-                isinstance(slice_value, tuple)
-                and slice_value.elts
-            )
-
-        self.set_precedence(ast._Precedence.ATOM, node.value)
-        self.traverse(node.value)
-        with self.delimit("[", "]"):
-            if is_non_empty_tuple(node.slice):
-                # parentheses can be omitted if the tuple isn't empty
-                self.items_view(self.traverse, node.slice.elts)
-            else:
-                self.traverse(node.slice)
-
-    def visit_Attribute(self, node):
-        self.set_precedence(ast._Precedence.ATOM, node.value)
-        self.traverse(node.value)
-        if isinstance(node.value, ast.Constant) and isinstance(node.value.value, int):
-            self.write(" ")
-        self.write(".")
-        self.write(node.attr)
-
-    def visit_Return(self, node):
-        self.fill("return")
-        if node.value:
-            self.write(" ")
-            self.traverse(node.value)
-
-
-class TuringTranslator(JuliaTranslator):
-    def probprog(self, name: str, args: list[str], body):
-        self.write("using Turing\n")
-        self.write("using Distributions\n")
-        self.write(f"@model function {name}(", ", ".join(args), ")")
-        with self.block():
-            self.traverse(body)
-
-    def probprog_sample(self, target, address, distribution_name, distribution_args, distribution_keywords):
-        self.fill()
-        self.set_precedence(ast._Precedence.TUPLE, target)
-        self.traverse(target)
-        self.write(" ~ ")
-        #self.traverse(address)
-        self.write(f"{distribution_name}")
-        self._write_arguments(distribution_args, distribution_keywords)
-
-    def probprog_observe(self, value, address, distribution_name, distribution_args, distribution_keywords):
-        self.traverse(value)
-        self.write(" ~ ")
-        # self.traverse(address)
-        self.write(f"{distribution_name}")
-        self._write_arguments(distribution_args, distribution_keywords)
-    
-    def probprog_boolean_observe(self, value, address):
-        raise NotImplementedError
-    
-    def probprog_factor(self, value):
-        raise NotImplementedError
-    
-class GenTranslator(JuliaTranslator):
-    def visit_Call(self, node):
-        match node:
-            case ast.Call(
-                func=ast.Attribute(value=ast.Name(id="pr"),attr="IndexedAddress"),
-                args = [ast.Constant(value=address), index]
-                ):
-                if isinstance(index, ast.Name):
-                    self.write(":" + address + " => " + index.id)
-                    return
-
-        super().visit_Call(node)
-
-    def probprog(self, name: str, args: list[str], body):
-        self.write("using Turing\n")
-        self.write("using Distributions\n")
-        self.write(f"@model function {name}(", ", ".join(args), ")")
-        with self.block():
-            self.traverse(body)
-
-    def probprog_sample(self, target, address, distribution_name, distribution_args, distribution_keywords):
-        self.fill()
-        self.set_precedence(ast._Precedence.TUPLE, target)
-        self.traverse(target)
-        self.write(" = ")
-        with self.delimit("{", "}"):
-            self.traverse(address)
-        self.write(" ~ ")
-        self.write(f"{distribution_name}")
-        self._write_arguments(distribution_args, distribution_keywords)
-
-    def probprog_observe(self, value, address, distribution_name, distribution_args, distribution_keywords):
-        with self.delimit("{", "}"):
-            self.traverse(address)
-        self.write(" ~ ")
-        self.write(f"{distribution_name}")
-        self._write_arguments(distribution_args, distribution_keywords)
-    
-    def probprog_boolean_observe(self, value, address):
-        raise NotImplementedError
-    
-    def probprog_factor(self, value):
-        raise NotImplementedError
-#%%
-import ast
-s = """
-@pr.probabilistic_program
-def coin_flips(data):
-    p = pr.sample("p", pr.Uniform(0, 1))
-    for i in range(len(data)):
-        pr.observe(data[i], pr.IndexedAddress("flip", i), pr.Bernoulli(p))
-    return p
-"""
-# %%
-a = ast.parse(s).body[0]
-print(ast.dump(a, indent=2))
-# %%
-translator = PyroTranslator()
-print(translator.visit(a))
-
-# %%
-translator = TuringTranslator()
-print(translator.visit(a))
-# %%
-translator = GenTranslator()
-print(translator.visit(a))
-# %%
